Refactor: divide ActionScreen's code(SelectItemPage)
